dynamic flask routing sample
this is rather pointless but it works well, and I think we should use url_for('function') instead of calling specific routes because it's a more dynamic base convention.

Also, this may come in handy when we have a form with multiple buttons, we can pass the variable as a part of directory path to access different files/functions based on which input is selected seamlessly via the one flask route.

Something to think about!