react with webpack utilized anywhere
so this actually doesn't work yet but also the bootstrap CSS is fucked up too and it takes 5 years to load so I think there might be something else going on? sus it out. The basic configuration of everything in the "react/" directory should be perfect. and it will drop a bundle in the static/js/reactBundles directory with a bundled js file ready to be used like any other js file.